Remove shebang line from genccaxml.
Fixed error in building cca-spec-babel with conda build 2 that
resulted in a bad interpreter error when running the genccaxml
Python script. This was caused by the long path name that conda
build was using for its build directory. It seems the shebang
line is truncated when run as a script.